refactored schedule_list_view out of schedule_page_view
